feat: update

+++ dev_iots_core_test/core/Globals.h
@@ -223,7 +223,7 @@ byte iotModuleState = IOT_MODULE_WAITING_INIT;
 byte iotConnState = IOT_CONN_WAITING_INIT;
 byte iotMqttMsgState = IOT_CONN_WAITING_INIT;
 
-byte iotParseIdx =0;
+int iotParseIdx = 0;
 
 String iotSerialRecv = "";
 String iotExtractedRecv = "";
